+++ docs/src/bfkit_comparison.jl
@@ -69,11 +69,11 @@ u0 = [-4.0, 5, 0] # state
 # Now we can create the bifurcation problem
 
 import BifurcationKit as BK
-using OrdinaryDiffEq
+using OrdinaryDiffEqVerner
 using CairoMakie
 
 bf_prob = BK.BifurcationProblem(
-    modified_lorenz_rule!, u0, p0, (BK.@lens _[pidx])
+    modified_lorenz_rule!, u0, p0, (BK.@optic _[pidx])
 )
 
 # and then a full solution structure from DifferentialEquations.jl, that
@@ -88,8 +88,8 @@ point_on_lc = [
     -5.178825669659272,
 ]
 
-ode_prob = ODEProblem(modified_lorenz_rule!, point_on_lc, (0.0, 100.0), p0)
-sol = OrdinaryDiffEq.solve(ode_prob; alg = Vern9(), abstol = 1e-9, reltol = 1e-9)
+ode_prob = ODEProblem(modified_lorenz_rule!, point_on_lc, (0.0, 50.0), p0)
+sol = OrdinaryDiffEqVerner.solve(ode_prob; alg = Vern9(), abstol = 1e-9, reltol = 1e-9)
 j = length(sol)÷2
 fig, ax = lines(sol.t[j:end], sol[1, j:end])
 lines!(ax, sol.t[j:end], sol[2, j:end])
@@ -105,15 +105,16 @@ fig
 # would need to fine tune to the problem at hand.
 opts_br = BK.ContinuationPar(
     p_min = prange[1], p_max = prange[end],
-    ds = 0.002, dsmax = 0.01, n_inversion = 6,
-    detect_bifurcation = 3, max_bisection_steps = 25, nev = 4,
+    ds = 0.002, dsmax = 0.01, dsmin = 1e-6, n_inversion = 6,
+    detect_bifurcation = 3, max_bisection_steps = 50, nev = 4,
     max_steps = 2000, tol_stability = 1e-3,
 )
 
 # We now create a periodic orbit problem type, by choosing a periodic
-# orbit finding algorithm
+# orbit finding algorithm. Here we will use an optimized Jacobian
+# for a collocation problem (advised via BifurcationKit.jl)
 
-periodic_orbit_algo = BK.PeriodicOrbitOCollProblem(40, 4)
+periodic_orbit_algo = BK.PeriodicOrbitOCollProblem(40, 4; jacobian = BK.DenseAnalyticalInplace())
 
 # and creating the problem type giving the period guess 19.0
 
@@ -124,7 +125,7 @@ probpo, cish = BK.generate_ci_problem(
 # To call the continuation we need to also tell it what aspects of the
 # periodic orbit to record, so we define
 
-argspo = (record_from_solution = (x, p) -> begin
+argspo = (record_from_solution = (x, p; k...) -> begin
 		xtt = BK.get_periodic_orbit(p.prob, x, p.p)
 		return (max = maximum(xtt[1,:]),
 				min = minimum(xtt[1,:]),
@@ -139,24 +140,39 @@ predictor = BK.PALC(tangent = BK.Bordered())
 # and _finally_ call the continuation from BK
 
 @time branch = BK.continuation(probpo, cish, predictor, opts_br;
-    verbosity = 0, plot = false, argspo...
+    verbosity = 0, plot = false,
+    linear_algo = BK.COPBLS(), # faster linear solver
+    argspo...,
+    bothside = true,
 )
 
-stability = branch.stable
-color = [s ? "black" : "red" for s in stability]
-marker = [s ? :circle : :x for s in stability]
-scatter(branch.branch.p, branch.branch.min; color, marker)
+# The converges within about 2 seconds, ignoring compilation time.
+# Let's plot the result:
 
-# The above code takes about 5 seconds to run.
-# Thankfully it works. Or rather, almost.
-# The code finds no stable limit cycle for parameter less than 5.0,
-# even though we know (see Tutorial final plot) that there is one.
-# Or maybe, it is an extremely weakly chaotic attractor with MLE almost 0.
-# Or maybe it is a quasiperiodic attractor. One needs to analyze further,
-# but Attractors.jl finds everything without much difficulty.
-# Altering the above code to start the continuation at 4.7 finds a limit cycle
-# there, but only manages to continue it only up to parameter 4.74,
-# instead of well into parameter 5.5 or more that Attractors.jl shows.
+stability = branch.stable
+fig, ax = scatter(
+    branch.branch.p[stability], branch.branch.min[stability];
+    label = "stable PO", color = "black", marker = :circle
+)
+scatter!(
+    branch.branch.p[.!stability], branch.branch.min[.!stability];
+    label = "unstable PO", color = "red", marker = :x
+
+)
+axislegend(ax)
+fig
+
+# Even the previous version that did work, did not find a
+# stable limit cycle for parameter less than 5.0,
+# In the main Tutorial we see that there is a limit cycle for
+# parameter value down to 4.7. Here we see that the limit cycle is actually
+# unstable. So the attractor found in the main Tutorial
+# could be a weakly chaotic attractor with MLE almost 0.
+# Or maybe it is a quasiperiodic attractor. Or maybe it is an alternative limit cycle
+# close to the one tracked here by BifurcationKit.jl.
+# One needs to analyze further!
+
+# %% #src
 
 # ## Attractors.jl version
 
@@ -174,8 +190,6 @@ scatter(branch.branch.p, branch.branch.min; color, marker)
 # 1. a global continuation algorithm, and optionally a matcher for it.
 # 1. a sampler to sample initial conditions in the state space.
 
-
-# %% #src
 using Attractors
 ds = CoupledODEs(modified_lorenz_rule!, u0, p0;
     diffeq = (alg = Vern9(), abstol = 1e-9, reltol = 1e-9)
@@ -211,7 +225,7 @@ plot_attractors_curves(
 # ## Discussion and comparison
 
 # Attractors.jl found not only a single limit cycle,
-# but also all system attractors, including chaotic ones.
+# but also all system attractors, including chaotic or quasiperiodic ones.
 # It didn't require any initial guess regarding the limit cycle or its period,
 # but only a state space box that may contain attractors.
 # Attractors.jl is extremely robust w.r.t. to its input parameters:
@@ -224,36 +238,36 @@ plot_attractors_curves(
 # And finally, Attractors.jl estimates a more general nonlocal measure of stability,
 # in the sense that if a set is nonlocally stable, it is guaranteed to be locally stable,
 # however the other way around isn't guaranteed.
-# Moreover, due to the orthogonality of finding and matching attractors,
-# as well as finding _all_ attractors, the global continuation of Attractors.jl
-# can continue along arbitrary user-defined curves in parameter space; not just
-# along a single parameter axis. This is possible because it is completely fine
-# for some attractors to stop existing during the global continuation,
-# while local continuation stops when attractors (and their unstable version)
-# stop existing.
+# The global continuation of Attractors.jl continues the whole of a
+# multistable state space across an arbitrary parameter curve.
+# It finds all attractors that exist at each parameter combination and
+# matches them to previous ones to establish continuity.
+# It is completely fine
+# for some attractors to stop existing during the global continuation.
 
-# Traditional local continuation can track the unstable
-# branches, and automatically detect and label local bifurcations,
-# both of which are not possible in Attractors.jl
-# (note we didn't bother to plot any of the detected bifurcations - if any are found).
+# Local continuation tracks a single and specific fixed point or limit cycle within a specified
+# parameter range. The continuation will stop if the object stops existing all-together.
+# But the local continuation can continue to track the object if it becomes unstable.
+# Local continuation also automatically detects and labels local bifurcations.
 # In our experience having the local bifurcations is always useful.
-# Now, whether the unstable branches are of a limit cycle are useful or not,
+# Now, whether the unstable branches of a limit cycle are useful or not,
 # depends on the research question and whether the analysis is done for some
 # sort of real world understanding (unstable limit cycles / fixed points don't
 # actually exist in the real world).
 # Beyond this however, BifurcationKit.jl is also optimised for PDE systems,
 # while Attractors.jl isn't.
 
-# Now, we have to be a bit more transparent here.
-# Besides this absence of stable orbits for parameter less than 5.0
-# that we discussed in the end of the BifurcationKit.jl version,
-# we need to say that it took **a lot of effort** to make the
-# BifurcationKit.jl code work. At least, a lot of effort compared with the effort
+# To be transparent, the biggest downside of local continuation software
+# (and not a particular downside of BifurcationKit.jl specifically),
+# is that it can take **a lot of effort** or **a lot of expertise** to make them work.
+# At least, a lot of effort compared with the effort
 # it took to make the Attractors.jl version work, which was simply "increase the recurrences
 # threshold", which is standard practice when dealing with chaotic systems
 # [Datseris2022](@cite). For example, any other of the
-# periodic orbit algorithms of BifurcationKit.jl (such as shooting or trapezoid) fails.
-# Using a slightly incorrect initial period guess of 20.0 instead of 19.0 also fails.
+# periodic orbit algorithms of BifurcationKit.jl (such as shooting or trapezoid) fails
+# unless its parameters are adjusted. Using alternative specifications for the Jacobian
+# or the linear solving algorithm can also lead to failure.
+# Using an incorrect initial period guess can also lead to failure.
 # We imagine that this sensitivity would apply also to some other of the
 # several meta-parameters that enter a traditional continuation routine,
 # for example the thresholds and accuracy parameters related to Newton convergence,