+++ ext/plotting.jl
@@ -33,10 +33,11 @@ end
 ##########################################################################################
 # Attractors
 ##########################################################################################
-function Attractors.plot_attractors(a; kw...)
+function Attractors.plot_attractors(a; access = SVector(1, 2), kw...)
     fig = Figure()
-    ax = Axis(fig[1,1])
-    plot_attractors!(ax, a; kw...)
+    AX = length(access) == 2 ? Axis : Axis3
+    ax = AX(fig[1,1])
+    plot_attractors!(ax, a; access, kw...)
     return fig
 end
 
@@ -52,8 +53,7 @@ function Attractors.plot_attractors!(ax, attractors;
     for k in ukeys
         k ∉ keys(attractors) && continue
         A = attractors[k]
-        x, y = columns(A[:, access])
-        scatter!(ax, x, y;
+        scatter!(ax, A[:, access];
             color = (colors[k], 0.9), markersize = 20,
             marker = markers[k],
             label = "$(labels[k])",
@@ -237,14 +237,14 @@ function Attractors.plot_basins_curves!(ax, fractions_cont, prange = 1:length(fr
     if style == :band
         # transform to cumulative sum
         for j in 2:length(bands)
-            bands[j] .+= bands[j-1]
+            bands[ukeys[j]] .+= bands[ukeys[j-1]]
         end
         for (j, k) in enumerate(ukeys)
             if j == 1
-                l, u = 0, bands[j]
+                l, u = 0, bands[k]
                 l = fill(0f0, length(u))
             else
-                l, u = bands[j-1], bands[j]
+                l, u = bands[ukeys[j-1]], bands[ukeys[j]]
             end
             band!(ax, prange, l, u;
                 color = colors[k], label = "$(labels[k])", series_kwargs...
@@ -255,8 +255,8 @@ function Attractors.plot_basins_curves!(ax, fractions_cont, prange = 1:length(fr
         end
         ylims!(ax, 0, 1)
     elseif style == :lines
-        for (j, k) in enumerate(ukeys)
-            scatterlines!(ax, prange, bands[j];
+        for k in ukeys
+            scatterlines!(ax, prange, bands[k];
                 color = colors[k], label = "$(labels[k])", marker = markers[k],
                 markersize = 10, linewidth = 3, series_kwargs...
             )
@@ -268,16 +268,6 @@ function Attractors.plot_basins_curves!(ax, fractions_cont, prange = 1:length(fr
     xlims!(ax, minimum(prange), maximum(prange))
     add_legend && axislegend(ax; axislegend_kwargs...)
     return
-end
-
-function continuation_series(continuation_info, defval, ukeys = unique_keys(continuation_info))
-    bands = [zeros(length(continuation_info)) for _ in ukeys]
-    for i in eachindex(continuation_info)
-        for (j, k) in enumerate(ukeys)
-            bands[j][i] = get(continuation_info[i], k, defval)
-        end
-    end
-    return bands
 end
 
 function Attractors.plot_attractors_curves(attractors_cont, attractor_to_real, prange = 1:length(attractors_cont); kwargs...)
@@ -310,21 +300,12 @@ function Attractors.plot_continuation_curves!(ax, continuation_info, prange = 1:
     )
 
     series = continuation_series(continuation_info, NaN, ukeys)
-    for (j, k) in enumerate(ukeys)
-        scatterlines!(ax, prange, series[j];
+    for k in ukeys
+        scatterlines!(ax, prange, series[k];
             color = colors[k], label = "$(labels[k])", marker = markers[k],
             markersize = 10, linewidth = 3, slines_kwargs...
         )
     end
-
-    # for i in eachindex(continuation_info)
-    #     info = continuation_info[i]
-    #     for (k, val) in info
-    #         scatter!(ax, prange[i], val;
-    #             color = colors[k], marker = markers[k], label = string(labels[k]),
-    #         )
-    #     end
-    # end
     xlims!(ax, minimum(prange), maximum(prange))
     add_legend && axislegend(ax; axislegend_kwargs..., unique = true)
     return